Migrate rate-limit module
