+++ packages/services/api/src/modules/rate-limit/resolvers.generated.ts
@@ -1,0 +1,7 @@
+/* This file was automatically generated. DO NOT UPDATE MANUALLY. */
+import type { Resolvers } from './../../__generated__/types.next';
+import { Organization } from './resolvers/Organization';
+
+export const resolvers: Resolvers = {
+  Organization: Organization,
+};
